refactor: moved topics in separate config
