+++ kafka_message_producer/src/main/java/ru/dankoy/kafkamessageproducer/config/kafka/KafkaWithProtobuf.java
@@ -14,11 +14,9 @@ import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.boot.ssl.SslBundles;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.config.TopicBuilder;
 import org.springframework.kafka.core.DefaultKafkaProducerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.core.ProducerFactory;
-import org.springframework.kafka.support.JacksonUtils;
 import ru.dankoy.kafkamessageproducer.core.service.messagesender.generic.KafkaTemplateGeneric;
 import ru.dankoy.kafkamessageproducer.core.service.messagesender.generic.KafkaTemplateGenericImpl;
 import ru.dankoy.kafkamessageproducer.core.service.messagesender.protobuf.MessageProducerServiceKafkaProtobuf;
@@ -30,32 +28,18 @@ public class KafkaWithProtobuf {
 
   private static final String HEADER_NAME = "subscription_type";
 
-  private final String coubCommunityTopicName;
-  private final String coubTagTopicName;
-  private final String coubChannelTopicName;
   private final String producerClientId;
   private final String schemaRegistryUrl;
 
   public KafkaWithProtobuf(
-      @Value("${application.kafka.topic.coub-com-subs}") String coubCommunityTopicName,
-      @Value("${application.kafka.topic.coub-tag-subs}") String coubTagTopicName,
-      @Value("${application.kafka.topic.coub-channel-subs}") String coubChannelTopicName,
       @Value("${application.kafka.producers.coubs.client-id}") String producerClientId,
       @Value("${application.kafka.schema-registry.url}") String schemaRegistryUrl) {
-    this.coubCommunityTopicName = coubCommunityTopicName;
-    this.coubTagTopicName = coubTagTopicName;
-    this.coubChannelTopicName = coubChannelTopicName;
     this.producerClientId = producerClientId;
     this.schemaRegistryUrl = schemaRegistryUrl;
   }
 
   @Bean
-  public ObjectMapper objectMapper() {
-    return JacksonUtils.enhancedObjectMapper();
-  }
-
-  @Bean
-  public ProducerFactory<String, Message> producerFactory(
+  public ProducerFactory<String, Message> protobufProducerFactory(
       KafkaProperties kafkaProperties, ObjectMapper mapper, SslBundles sslBundles) {
     var props = kafkaProperties.buildProducerProperties(sslBundles);
     props.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, "true");
@@ -75,60 +59,44 @@ public class KafkaWithProtobuf {
   }
 
   @Bean
-  public KafkaTemplate<String, Message> kafkaTemplate(
-      ProducerFactory<String, Message> producerFactory) {
-    return new KafkaTemplate<>(producerFactory);
+  public KafkaTemplate<String, Message> kafkaTemplateProtobufProducer(
+      ProducerFactory<String, Message> protobufProducerFactory) {
+    return new KafkaTemplate<>(protobufProducerFactory);
   }
 
   @Bean
-  public NewTopic topicCoubCommunityMessage() {
-    return TopicBuilder.name(coubCommunityTopicName).partitions(2).replicas(1).build();
+  public KafkaTemplateGeneric<String, Message> kafkaTemplateProtobuf(
+      KafkaTemplate<String, Message> kafkaTemplateProtobufProducer) {
+
+    return new KafkaTemplateGenericImpl<>(kafkaTemplateProtobufProducer);
   }
 
   @Bean
-  public NewTopic topicCoubTagMessage() {
-    return TopicBuilder.name(coubTagTopicName).partitions(2).replicas(1).build();
-  }
-
-  @Bean
-  public NewTopic topicCoubChannelMessage() {
-    return TopicBuilder.name(coubChannelTopicName).partitions(2).replicas(1).build();
-  }
-
-  @Bean
-  public KafkaTemplateGeneric<String, Message> kafkaTemplateCoubMessage(
-      KafkaTemplate<String, Message> kafkaTemplate) {
-
-    return new KafkaTemplateGenericImpl<>(kafkaTemplate);
-  }
-
-  @Bean
-  public MessageProducerServiceKafkaProtobuf channelMessageProducerServiceKafka(
+  public MessageProducerServiceKafkaProtobuf channelMessageProducerServiceKafkaProtobuf(
       NewTopic topicCoubChannelMessage,
-      KafkaTemplateGeneric<String, Message> kafkaTemplateCoubMessage) {
+      KafkaTemplateGeneric<String, Message> kafkaTemplateProtobuf) {
     return new MessageProducerServiceKafkaProtobufImpl(
         topicCoubChannelMessage.name(),
-        kafkaTemplateCoubMessage,
+        kafkaTemplateProtobuf,
         r -> r.headers().add(HEADER_NAME, "BY_CHANNEL".getBytes(StandardCharsets.UTF_8)));
   }
 
   @Bean
-  public MessageProducerServiceKafkaProtobuf communityMessageProducerServiceKafka(
+  public MessageProducerServiceKafkaProtobuf communityMessageProducerServiceKafkaProtobuf(
       NewTopic topicCoubCommunityMessage,
-      KafkaTemplateGeneric<String, Message> kafkaTemplateCoubMessage) {
+      KafkaTemplateGeneric<String, Message> kafkaTemplateProtobuf) {
     return new MessageProducerServiceKafkaProtobufImpl(
         topicCoubCommunityMessage.name(),
-        kafkaTemplateCoubMessage,
+        kafkaTemplateProtobuf,
         r -> r.headers().add(HEADER_NAME, "BY_COMMUNITY".getBytes(StandardCharsets.UTF_8)));
   }
 
   @Bean
-  public MessageProducerServiceKafkaProtobuf tagMessageProducerServiceKafka(
-      NewTopic topicCoubTagMessage,
-      KafkaTemplateGeneric<String, Message> kafkaTemplateCoubMessage) {
+  public MessageProducerServiceKafkaProtobuf tagMessageProducerServiceKafkaProtobuf(
+      NewTopic topicCoubTagMessage, KafkaTemplateGeneric<String, Message> kafkaTemplateProtobuf) {
     return new MessageProducerServiceKafkaProtobufImpl(
         topicCoubTagMessage.name(),
-        kafkaTemplateCoubMessage,
+        kafkaTemplateProtobuf,
         r -> r.headers().add(HEADER_NAME, "BY_TAG".getBytes(StandardCharsets.UTF_8)));
   }
 }
